feat/refactor: Feedback and FeedbackPopup components
Separates the feedback form to a `<Feedback>` component, adds a
`<FeedbackPopup>` component for displaying a less intrusive feedback
form for requesting user feedback automatically and refactors the
`<FeedbackModal>` component.
