+++ frontend/src/lib/components/feedback/index.ts
@@ -1,0 +1,2 @@
+export {default as Feedback} from './Feedback.svelte';
+export * from './Feedback.type';
