+++ frontend/src/lib/components/feedback/popup/FeedbackPopup.type.ts
@@ -1,0 +1,3 @@
+import type {AlertProps} from '$lib/components/alert';
+
+export type FeedbackPopupProps = Partial<AlertProps>;
