+++ frontend/src/lib/components/feedback/popup/index.ts
@@ -1,0 +1,2 @@
+export {default as FeedbackPopup} from './FeedbackPopup.svelte';
+export * from './FeedbackPopup.type';
