+++ frontend/src/lib/i18n/translations/fi/feedback.json
@@ -1,10 +1,11 @@
 {
   "descriptionLabel": "Muut kommentit",
-  "descriptionPlaceholder": "Kirjoita muut kommenttisi tai ehdotuksesi tähän",
+  "descriptionPlaceholder": "Muut kommentit tai ehdotukset",
   "emailIntro": "Voit halutessasi myös lähettää meille sähköpostia osoitteeseen",
   "error": "Anteeksi, palautteen lähettämisessä tapahtui virhe.",
   "errorEmailIntro": "Voisitko ystävällisesti lähettää palautteesi sähköpostilla klikkaamalla alta. Jo kirjoittamasi viesti kopioidaan automaattisesti mukaan.",
   "errorEmailSubject": "Palaute",
+  "popupTitle": "Mitä mieltä olet vaalikoneesta?",
   "ratingLabel": "Yleisarvio vaalikoneesta",
   "ratingValueLabel": "{rating, plural, =0 {Ei arviota} other {Arvio # maksimista {ratingMax}}}",
   "sendButton": "Lähetä palaute",
